requirements.txt
langchain-core
langchain-openai
python-dotenv
pydantic
fastapi
uvicorn
